requirements.txt
loguru==0.7.3

numpy==2.2.6

# Image manipulation
opencv-contrib-python-headless==4.12.0.88
Pillow==11.3.0
ultralytics==8.3.170

# Asynchronous HTTP requests
aiohttp==3.12.14
httpx==0.28.1
brotli==1.1.0
selenium==4.34.2
selenium-stealth==1.0.6
urllib3==2.5.0

# Connecting to the PostgreSQL database
asyncpg==0.30.0

beautifulsoup4==4.13.4
soupsieve==2.7

# Testing differences in dictionaries
deepdiff==8.5.0

# Used with the Figure Separator
pycocotools==2.0.10

pytz==2025.2
typing_extensions==4.14.0
PyYAML==6.0.2
pymupdf==1.26.3

torch==2.7.1
torchvision==0.22.1
pytorch-model-summary==0.1.2

# Run LLMs
ollama==0.5.1
openai==1.97.1
asksageclient==1.31

# Used with the API
dash[async]==3.1.1
fastapi[standard]==0.116.1
pydantic==2.11.7
starlette==0.46.2
uvicorn==0.35.0
hypercorn==0.17.3
sqlmodel==0.0.24
sqlalchemy==2.0.41
pydantic_settings==2.10.1
uuid-utils==0.11.0

# Used with the Dashboard
dash_bootstrap_components==2.0.3
gunicorn[gevent]==23.0.0
certifi==2025.7.14
